requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
h11==0.12.0
idna==3.2
motor==2.4.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
